cek migration quantity yang blon FLOAT [MIGRATE FRESH !!!]

+++ database/migrations/2018_10_24_023853_create_trx_goods_return_detail_table.php
@@ -16,7 +16,7 @@ class CreateTrxGoodsReturnDetailTable extends Migration
         Schema::create('trx_goods_return_detail', function (Blueprint $table) {
             $table->increments('id');
             $table->unsignedInteger('goods_return_id');
-            $table->integer('quantity');
+            $table->float('quantity',15,2);
             $table->unsignedInteger('material_id')->nullable();
             $table->unsignedInteger('resource_detail_id')->nullable();
             $table->unsignedInteger('storage_location_id')->nullable();
